Restructure to return early

+++ src/sphinxcli/clean.py
@@ -5,28 +5,36 @@ from shutil import rmtree
 import rich
 
 
-def clean_docs(target: Path, clean_all: bool = False) -> None:
-    if not clean_all:
-        rich.print(f"[green]Removing documents in {target}...[/]")
-
+def clean_docs(target: Path) -> bool:
     target = target.resolve()
-    for item in target.glob("*"):
+    targets = list(target.glob("*"))
+
+    if len(targets) == 0:
+        return False
+
+    for item in targets:
         if item.is_dir():
             rmtree(item)
         else:
             item.unlink()
 
+    return True
 
-def clean_doctrees(doctrees: Path, clean_all: bool = False) -> None:
-    if doctrees.exists():
-        if not clean_all:
-            rich.print(f"[green]Removing doctrees in {doctrees}...[/]")
+
+def clean_doctrees(doctrees: Path) -> bool:
+    if not doctrees.exists():
+        return False
+    else:
         rmtree(doctrees)
+        return True
 
 
 def clean_all(target: Path, doctrees: Path):
     rich.print("[green]Cleaning all generated files[/]")
-    if commonpath([doctrees, target]) != target:
-        clean_doctrees(doctrees)
 
-    clean_docs(target)
+    if clean_docs(target):
+        rich.print(f"  [green]Removed documents in {target}...[/]")
+
+    if commonpath([doctrees, target]) != target:
+        if clean_doctrees(doctrees):
+            rich.print(f"  [green]Removed doctrees in {doctrees}...[/]")
